+++ math/tuples.c
@@ -6,13 +6,13 @@
 /*   By: bismail <bismail@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/10 14:29:01 by abueskander       #+#    #+#             */
-/*   Updated: 2025/02/12 10:45:09 by bismail          ###   ########.fr       */
+/*   Updated: 2025/02/12 11:54:00 by bismail          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <object.h>
 
-t_tuple *point(double x, double y, double z)
+t_tuple *point(float x, float y, float z)
 {
         t_tuple *pnt;
 
@@ -25,7 +25,7 @@ t_tuple *point(double x, double y, double z)
         pnt->W = POINT;
         return pnt;
 }
-t_tuple *vector(double x, double y, double z)
+t_tuple *vector(float x, float y, float z)
 {
         t_tuple *vec;
